change: replace typescript with tsd for type testing
