--- tsconfig.json
@@ -1,34 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "esnext",
-    "noEmit": true,
-    "strict": true,
-    "noImplicitAny": true,
-    "esModuleInterop": true,
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "incremental": true
-  },
-  "files": [
-    "./types.test.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ],
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
-  ]
-}
